feat: Get usage on self-hosted instance > WIP
