+++ backend/api/src/main/kotlin/io/tolgee/hateoas/ee/SelfHostedEeSubscriptionModel.kt
@@ -16,6 +16,7 @@ import java.math.BigDecimal
 @Relation(collectionRelation = "subscriptions", itemRelation = "subscription")
 open class SelfHostedEeSubscriptionModel(
   val id: Long = 0,
+  val cancelAtPeriodEnd: Boolean = false,
   val currentPeriodStart: Long? = null,
   val currentPeriodEnd: Long? = null,
   val currentBillingPeriod: BillingPeriod = BillingPeriod.MONTHLY,
