+++ backend/data/src/main/kotlin/io/tolgee/constants/Message.kt
@@ -278,6 +278,7 @@ enum class Message {
   PLAN_KEY_LIMIT_EXCEEDED,
   KEYS_SPENDING_LIMIT_EXCEEDED,
   PLAN_SEAT_LIMIT_EXCEEDED,
+  INSTANCE_NOT_USING_LICENSE_KEY,
   ;
 
   val code: String
