+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/api/v2/controllers/EeSubscriptionUsageController.kt
@@ -1,0 +1,33 @@
+package io.tolgee.ee.api.v2.controllers
+
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.tags.Tag
+import io.tolgee.ee.api.v2.hateoas.assemblers.EeSubscriptionModelAssembler
+import io.tolgee.ee.service.eeSubscription.EeSubscriptionServiceImpl
+import io.tolgee.ee.service.eeSubscription.EeSubscriptionUsageService
+import io.tolgee.hateoas.ee.uasge.UsageModel
+import io.tolgee.openApiDocs.OpenApiEeExtension
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+/**
+ *
+ * It returns info about the current ee subscription on self-hosted instance.
+ *
+ * This controller is relevant only for self-hosted instances.
+ * It doesn't make sense on Tolgee Cloud.
+ */
+@RestController
+@RequestMapping("/v2/ee-subscription-usage/")
+@Tag(name = "EE Licence")
+@OpenApiEeExtension
+class EeSubscriptionUsageController(
+  private val usageService: EeSubscriptionUsageService
+) {
+  @GetMapping("")
+  @Operation(summary = "Get the info about the current EE subscription")
+  fun getUsage(): UsageModel? {
+    return usageService.getUsage()
+  }
+}
