+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/api/v2/controllers/internal/BypassSeatCountCheckController.kt
@@ -2,7 +2,7 @@ package io.tolgee.controllers.internal
 
 import io.swagger.v3.oas.annotations.Hidden
 import io.tolgee.configuration.tolgee.TolgeeProperties
-import io.tolgee.ee.service.EeSubscriptionServiceImpl
+import io.tolgee.ee.service.eeSubscription.EeSubscriptionServiceImpl
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.bind.annotation.CrossOrigin
 import org.springframework.web.bind.annotation.PutMapping
