+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/component/EeSubscriptionInfoProviderImpl.kt
@@ -1,7 +1,7 @@
 package io.tolgee.ee.component
 
 import io.tolgee.component.EeSubscriptionInfoProvider
-import io.tolgee.ee.service.EeSubscriptionServiceImpl
+import io.tolgee.ee.service.eeSubscription.EeSubscriptionServiceImpl
 import org.springframework.stereotype.Component
 
 @Component
