+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/component/contentDelivery/EeTolgeeTranslateApiServiceImpl.kt
@@ -7,7 +7,8 @@ import io.tolgee.component.machineTranslation.providers.tolgee.TolgeeTranslateAp
 import io.tolgee.component.machineTranslation.providers.tolgee.TolgeeTranslateParams
 import io.tolgee.configuration.tolgee.TolgeeProperties
 import io.tolgee.constants.Message
-import io.tolgee.ee.service.EeSubscriptionServiceImpl
+import io.tolgee.ee.service.eeSubscription.EeSubscriptionErrorCatchingService
+import io.tolgee.ee.service.eeSubscription.EeSubscriptionServiceImpl
 import io.tolgee.exceptions.BadRequestException
 import io.tolgee.exceptions.OutOfCreditsException
 import org.springframework.http.HttpHeaders
@@ -21,6 +22,7 @@ class EeTolgeeTranslateApiServiceImpl(
   private val httpClient: HttpClient,
   private val subscriptionService: EeSubscriptionServiceImpl,
   private val eeSubscriptionServiceImpl: EeSubscriptionServiceImpl,
+  private val catchingService: EeSubscriptionErrorCatchingService
 ) : TolgeeTranslateApiService, EeTolgeeTranslateApiService {
   companion object {
     const val API_PATH = "v2/public/translator/translate"
@@ -32,7 +34,7 @@ class EeTolgeeTranslateApiServiceImpl(
       subscriptionService.findSubscriptionDto()?.licenseKey ?: throw IllegalStateException("Not Subscribed")
 
     try {
-      return subscriptionService.catchingLicenseNotFound {
+      return catchingService.catchingLicenseNotFound {
         httpClient.requestForJson(
           url = url,
           body = params,
