+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/data/GetEeSubscriptionUsageDto.kt
@@ -1,0 +1,8 @@
+package io.tolgee.ee.data
+
+import jakarta.validation.constraints.NotBlank
+
+class GetEeSubscriptionUsageDto(
+  @field:NotBlank
+  var licenseKey: String
+)
