+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/service/eeSubscription/EeSubscriptionUsageService.kt
@@ -1,0 +1,26 @@
+package io.tolgee.ee.service.eeSubscription
+
+import io.tolgee.constants.Message
+import io.tolgee.ee.service.eeSubscription.cloudClient.TolgeeCloudLicencingClient
+import io.tolgee.exceptions.NotFoundException
+import io.tolgee.hateoas.ee.uasge.UsageModel
+import org.springframework.context.annotation.Lazy
+import org.springframework.stereotype.Service
+
+@Service
+class EeSubscriptionUsageService(
+  private val client: TolgeeCloudLicencingClient,
+  @Lazy
+  private val eeSubscriptionService: EeSubscriptionServiceImpl,
+  private val catchingService: EeSubscriptionErrorCatchingService
+) {
+  fun getUsage(): UsageModel {
+    val eeSubscription =
+      eeSubscriptionService.findSubscriptionDto()
+        ?: throw NotFoundException(Message.INSTANCE_NOT_USING_LICENSE_KEY)
+
+    return catchingService.catchingLicenseNotFound {
+      client.getUsageRemote(eeSubscription.licenseKey)
+    }
+  }
+}
