+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/service/eeSubscription/cloudClient/exceptions.kt
@@ -1,0 +1,5 @@
+package io.tolgee.ee.service.eeSubscription.cloudClient
+
+import io.tolgee.exceptions.NotFoundException
+
+class LicenseKeyNotFoundException : NotFoundException()
