+++ webapp/src/ee/billing/limitPopover/SpendingLimitExceeded.tsx
@@ -13,6 +13,7 @@ import { useOrganizationUsage } from 'tg.globalContext/helpers';
 import { USAGE_ELEMENT_ID } from '../component/CriticalUsageCircle';
 import { SpendingLimitExceededDescription } from 'tg.component/security/SignUp/SpendingLimitExceededDesciption';
 import { getProgressData } from '../component/getProgressData';
+import { PlanLimitPopoverWrapper } from './generic/PlanLimitPopoverWrapper';
 
 const StyledDialogContent = styled(DialogContent)`
   display: grid;
@@ -29,28 +30,11 @@ export const SpendingLimitExceededPopover: React.FC<Props> = ({
   open,
   onClose,
 }) => {
-  const { usage } = useOrganizationUsage();
-
-  const anchorEl = document.getElementById(USAGE_ELEMENT_ID);
-  const progressData = usage && getProgressData({ usage });
-
-  return progressData ? (
-    <Popover
+  return (
+    <PlanLimitPopoverWrapper
       open={open}
       onClose={onClose}
-      anchorEl={open ? anchorEl : undefined}
-      aria-labelledby="alert-dialog-title"
-      aria-describedby="alert-dialog-description"
       data-cy="spending-limit-exceeded-popover"
-      {...(anchorEl
-        ? {
-            anchorOrigin: { horizontal: 'right', vertical: 'bottom' },
-            transformOrigin: { horizontal: 'right', vertical: 'top' },
-          }
-        : {
-            anchorOrigin: { horizontal: 'center', vertical: 'top' },
-            transformOrigin: { horizontal: 'center', vertical: 'center' },
-          })}
     >
       <DialogTitle id="alert-dialog-title">
         {<T keyName="spending_limit_dialog_title" />}
@@ -71,6 +55,6 @@ export const SpendingLimitExceededPopover: React.FC<Props> = ({
           <T keyName="spending_limit_dialog_close" />
         </Button>
       </DialogActions>
-    </Popover>
-  ) : null;
+    </PlanLimitPopoverWrapper>
+  );
 };
